# wallbox/wallbox_mobile/API-DOC/swagger2.yml
openapi: 3.0.1
info:
  title: Wallbox Project API
  description: API for Wallbox mobile applications (Odoo v18 backend)
  version: "1.1.0"
  contact:
    name: API Support
    email: munafkhan732@gmail.com
servers:
  - url: https://qariqa.qariqa.com/v1
    description: Production server
  - url: https://qariqa-staging.qariqa.com/v1
    description: Staging server
...
paths:
...
  /my/wallbox/charging_stations:
    get:
      tags: [Charging Stations]
      summary: Get My Charging Stations
      description: Returns all charging stations associated with the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /my/wallbox/charging_stations/{station_id}:
    parameters:
      - $ref: '#/components/parameters/StationId'
    get:
      tags: [Charging Stations]
      summary: Get Charging Station Details
      description: Returns detailed information for a specific charging station.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags: [Charging Stations]
      summary: Update Charging Station
      description: Updates editable information for a specific charging station, like price or location.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargingStationUpdateRequest'
      responses:
        '200':
          description: Station updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /my/wallbox/charging_stations/{station_id}/sessions:
    parameters:
      - $ref: '#/components/parameters/StationId'
    get:
      tags: [Charging Stations]
      summary: Get Historical Sessions
      description: Returns all historical (completed/failed) charging sessions for a specific station.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingSessionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /my/wallbox/charging_stations/{station_id}/sessions/ongoing:
    parameters:
      - $ref: '#/components/parameters/StationId'
    get:
      tags: [Charging Stations]
      summary: Get Ongoing Sessions
      description: Returns ongoing charging sessions for a specific station.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingSessionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    ChargingStation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        building:
          type: string
        parking_space:
          type: string
        installation_date:
          type: string
        location:
          type: string
        charging_power:
          type: number
        guest_max_amount_limit:
          type: number
        connector_type:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        number_of_ports:
          type: integer
        owner_id:
          type: integer
        owner_name:
          type: string
        access_type:
          type: string
        authentication_method:
          type: string
        energy_source:
          type: string
        billing_system:
          type: string
        last_maintenance_date:
          type: string
        next_scheduled_maintenance:
          type: string
        reported_issues:
          type: string
        wallbox_order_id:
          type: integer
        number_of_charging_sessions:
          type: integer
        total_allowed_users:
          type: integer
        total_rfid_tags:
          type: integer
        charger_id:
          type: string
        price_per_kwh:
          type: number
        status:
          type: string
        ws_url:
          type: string
        total_energy:
          type: number
        total_recharged_cost:
          type: number
    ChargingStationUpdateRequest:
      type: object
      properties:
        guest_max_amount_limit:
          type: number
          example: 50
        latitude:
          type: number
          example: 40.7128
        longitude:
          type: number
          example: -74.0060
        price_per_kwh:
          type: number
          example: 0.25
    ChargingStationResponse:
      type: object
      properties:
        status:
          type: boolean
        charging_station:
          $ref: '#/components/schemas/ChargingStation'
    ChargingStationListResponse:
      type: object
      properties:
        status:
          type: boolean
        charging_stations:
          type: array
          items:
            $ref: '#/components/schemas/ChargingStation'
        count:
          type: integer

    # RFID Tag Schemas
...
    ChargingSession:
      type: object
      properties:
        id:
          type: integer
        transaction_id:
          type: string
        charging_station_id:
          type: string
        customer_id:
          type: string
        max_amount_limit:
          type: number
        start_meter:
          type: number
        stop_meter:
          type: number
        total_energy:
          type: number
        cost:
          type: number
        status:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        start_time:
          type: string
          nullable: true
        end_time:
          type: string
          nullable: true
        total_duration:
          type: string
        request_id:
          type: string
    ChargingSessionListResponse:
      type: object
      properties:
        status:
          type: boolean
        charging_sessions:
          type: array
          items:
            $ref: '#/components/schemas/ChargingSession'
        count:
          type: integer

    # Invitation Schemas
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              nullable: true
      required: [status, error]
  parameters:
...
    StationId:
      name: station_id
      in: path
      required: true
      schema:
        type: integer
      description: Charging Station ID
...
  responses:
    BadRequest:
      description: Bad request (invalid input)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "VALIDATION_ERROR"
              message: "Required fields are missing"
    Unauthorized:
      description: Unauthorized (invalid or missing token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "TOKEN_INVALID"
              message: "Invalid or expired authentication token"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "RESOURCE_NOT_FOUND"
              message: "The requested resource was not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT